Separate statistics from personal info reports Fixes https://jira.magfest.net/browse/MAGDEV-680. All non-personal data reports now belong in 'statistics' (with some exceptions, e.g., merch reports)
