+++ uber/templates/statistics/badges_sold.html
