+++ uber/templates/statistics/map.html
